backend/agents/personal_health_coach/pyproject.toml
[project]
name = "coral-rag-agent"
version = "0.1.0"
description = "Agent that uses RAG for memory retrieval"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai==0.126.0",
    "python-dotenv",
    "crewai-tools==0.47.1",
    "crewai-tools[mcp]",
    "docx2txt>=0.9",
]
